requirements.txt
pyTelegramBotAPI==4.16.1
elevenlabs==1.6.0
python-dotenv==1.0.1
